requirements.txt
streamlit>=1.28.0
arxiv>=1.4.0
requests>=2.31.0
sendgrid>=6.10.0
python-dotenv>=1.0.0
urllib3>=1.26.0